+++ Mepas.Sub.API/DAL/Repositories/ISubscriberRepository.cs
