+++ Mepas.Sub.API/Helpers/DateHelper.cs
@@ -1,25 +1,14 @@
-﻿namespace Mepas.Sub.API.Helpers
+﻿public static class DateHelper
 {
-    public static class DateHelper
+    public static (DateTime, DateTime) ParseDateRange(string startDate, string endDate)
     {
-        // Tarih aralığının geçerliliğini kontrol eder
-        public static bool IsValidDateRange(DateTime? startDate, DateTime? endDate)
-        {
-            // Eğer başlangıç veya bitiş tarihi boşsa geçerli kabul edilir
-            if (!startDate.HasValue || !endDate.HasValue)
-            {
-                return true;
-            }
-            return startDate <= endDate;
-        }
+        DateTime start = DateTime.Parse(startDate);
+        DateTime end = DateTime.Parse(endDate);
+        return (start, end);
+    }
 
-        // Tarihleri parse eder
-        public static (DateTime? StartDate, DateTime? EndDate) ParseDateRange(string startDate, string endDate)
-        {
-            DateTime? start = string.IsNullOrEmpty(startDate) ? null : DateTime.Parse(startDate);
-            DateTime? end = string.IsNullOrEmpty(endDate) ? null : DateTime.Parse(endDate);
-
-            return (start, end);
-        }
+    public static bool IsValidDateRange(DateTime startDate, DateTime endDate)
+    {
+        return startDate <= endDate;
     }
 }
